+++ packages/@expo/cli/src/export/exportAsync.ts
@@ -8,10 +8,16 @@ import { waitUntilAtlasExportIsReadyAsync } from '../start/server/metro/debuggin
 import { FileNotifier } from '../utils/FileNotifier';
 import { ensureDirectoryAsync, removeAsync } from '../utils/dir';
 import { ensureProcessExitsAfterDelay } from '../utils/exit';
+import { CommandError } from '../utils/errors';
 
 export async function exportAsync(projectRoot: string, options: Options) {
   // Ensure the output directory is created
   const outputPath = path.resolve(projectRoot, options.outputDir);
+
+  if (outputPath === projectRoot) {
+    throw new CommandError('--output-dir cannot be the same as the project directory.');
+  }
+
   // Delete the output directory if it exists
   await removeAsync(outputPath);
   // Create the output directory